+++ client/src/components/ui/label.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import * as LabelPrimitive from "@radix-ui/react-label";
+import { Label as LabelPrimitive } from "radix-ui";
 
 import { cn, getTextColor } from "../../lib/utils";
 import { useSettings } from "@/pages/settings/SettingsContext";
